#278 change datalabel, tooltip in bar, line chart
