+++ discovery-frontend/src/app/common/component/chart/option/converter/format-option-converter.ts
@@ -309,14 +309,14 @@ export class FormatOptionConverter {
 
           let pivotList = _.cloneDeep(pivot.rows);
           // category Name List 설정
-          result = this.getTooltipName(seriesNameList, pivotList, result, true, pivot, ShelveType.ROWS);
+          result = this.getTooltipName(seriesNameList, pivotList, result, false, pivot, ShelveType.ROWS);
 
           // aggregations의 measure가 2개이상인경우
         } else if (pivot.aggregations.length > 1) {
 
           let pivotList = _.cloneDeep(pivot.aggregations);
           // category Name List 설정
-          result = this.getTooltipName(seriesNameList, pivotList, result, true, pivot, ShelveType.AGGREGATIONS);
+          result = this.getTooltipName(seriesNameList, pivotList, result, false, pivot, ShelveType.AGGREGATIONS);
 
           // 단일시리즈인 경우 category Name으로 설정
         } else {
@@ -629,7 +629,7 @@ export class FormatOptionConverter {
     if( uiData ) {
 
       if (!uiOption.toolTip) uiOption.toolTip = {};
-      if (!uiOption.toolTip.displayTypes) uiOption.toolTip.displayTypes = this.setDisplayTypes(uiOption.type);
+      if (!uiOption.toolTip.displayTypes) uiOption.toolTip.displayTypes = this.setDisplayTypes(uiOption.type, pivot);
 
       // UI 데이터 가공
       let result: string[] = [];
@@ -770,7 +770,7 @@ export class FormatOptionConverter {
   /**
    * 차트별 displayTypes 기본값 설정
    */
-  public static setDisplayTypes(chartType: ChartType): UIChartDataLabelDisplayType[] {
+  public static setDisplayTypes(chartType: ChartType, pivot?: Pivot): UIChartDataLabelDisplayType[] {
 
     let displayTypes = [];
 
@@ -778,6 +778,18 @@ export class FormatOptionConverter {
 
       case ChartType.BAR:
       case ChartType.LINE:
+        // when bar, line chart has single series
+        if ((chartType === ChartType.BAR && pivot.aggregations.length <= 1 && pivot.rows.length < 1) ||
+          (chartType === ChartType.LINE && pivot.aggregations.length <= 1)) {
+          displayTypes[0] = UIChartDataLabelDisplayType.CATEGORY_NAME;
+          displayTypes[1] = UIChartDataLabelDisplayType.CATEGORY_VALUE;
+          // when bar, line chart has multi series
+        } else {
+          displayTypes[3] = UIChartDataLabelDisplayType.SERIES_NAME;
+          displayTypes[4] = UIChartDataLabelDisplayType.SERIES_VALUE;
+        }
+        break;
+
       case ChartType.CONTROL:
       case ChartType.COMBINE:
       case ChartType.WATERFALL:
