+++ discovery-frontend/src/app/common/component/chart/option/converter/label-option-converter.ts
@@ -17,7 +17,9 @@ import { UIOption } from '../ui-option';
 import * as _ from 'lodash';
 import { Series } from '../define/series';
 import { UIChartDataLabel } from '../ui-option/ui-datalabel';
-import { DataLabelPosition, Position } from '../define/common';
+import { DataLabelPosition, Position, UIChartDataLabelDisplayType } from '../define/common';
+import { FormatOptionConverter } from './format-option-converter';
+import { UIChartFormat } from '../ui-option/ui-format';
 
 /**
  * 데이터 레이블 옵션 컨버터
@@ -390,6 +392,73 @@ export class LabelOptionConverter {
 
     // 반환
     return chartOption;
+  }
+
+
+  /**
+   * set datalabel previewlist
+   */
+  public static setDataLabelPreviewList(uiOption: UIOption): Object[] {
+
+    // 미리보기 리스트 초기화
+    uiOption.dataLabel.previewList = [];
+
+    let format: UIChartFormat = uiOption.valueFormat;
+
+    // 축의 포멧이 있는경우 축의 포멧으로 설정
+    const axisFormat = FormatOptionConverter.getlabelAxisScaleFormat(uiOption);
+    if (axisFormat) format = axisFormat;
+
+    // 포멧값이 설정된 숫자값
+    let numValue = FormatOptionConverter.getFormatValue(1000, format);
+
+    if (uiOption.dataLabel.displayTypes) {
+      // displayType에 따라서 미리보기 설정
+      for (const type of uiOption.dataLabel.displayTypes) {
+
+        switch(type) {
+
+          case UIChartDataLabelDisplayType.CATEGORY_NAME:
+            uiOption.dataLabel.previewList.push({name: 'Category Name', value: UIChartDataLabelDisplayType.CATEGORY_NAME});
+            break;
+          case UIChartDataLabelDisplayType.CATEGORY_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.CATEGORY_VALUE});
+            break;
+          case UIChartDataLabelDisplayType.CATEGORY_PERCENT:
+            uiOption.dataLabel.previewList.push({name: '100%', value: UIChartDataLabelDisplayType.CATEGORY_PERCENT});
+            break;
+          case UIChartDataLabelDisplayType.SERIES_NAME:
+            uiOption.dataLabel.previewList.push({name: 'Series Name', value: UIChartDataLabelDisplayType.SERIES_NAME});
+            break;
+          case UIChartDataLabelDisplayType.SERIES_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.SERIES_VALUE});
+            break;
+          case UIChartDataLabelDisplayType.SERIES_PERCENT:
+            uiOption.dataLabel.previewList.push({name: '100%', value: UIChartDataLabelDisplayType.SERIES_PERCENT});
+            break;
+          case UIChartDataLabelDisplayType.XAXIS_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.XAXIS_VALUE});
+            break;
+          case UIChartDataLabelDisplayType.YAXIS_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.YAXIS_VALUE});
+            break;
+          case UIChartDataLabelDisplayType.VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.VALUE});
+            break;
+          case UIChartDataLabelDisplayType.NODE_NAME:
+            uiOption.dataLabel.previewList.push({name: 'Node Name', value: UIChartDataLabelDisplayType.NODE_NAME});
+            break;
+          case UIChartDataLabelDisplayType.LINK_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.LINK_VALUE});
+            break;
+          case UIChartDataLabelDisplayType.NODE_VALUE:
+            uiOption.dataLabel.previewList.push({name: numValue, value: UIChartDataLabelDisplayType.NODE_VALUE});
+            break;
+        }
+      }
+    }
+
+    return uiOption.dataLabel.previewList;
   }
   /*-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
    | Private Method
