+++ discovery-frontend/src/app/common/component/chart/type/line-chart/line-chart.component.ts
@@ -23,7 +23,7 @@ import {
   ChartColorList,
   ChartColorType,
   ChartType,
-  ColorCustomMode,
+  ColorCustomMode, EventType,
   LineCornerType,
   LineStyle,
   LineType,
@@ -56,6 +56,8 @@ import {UIChartAxis, UIChartAxisGrid, UIChartAxisLabelValue} from "../../option/
 import {AxisOptionConverter} from "../../option/converter/axis-option-converter";
 import {Axis as AxisDefine} from "../../option/define/axis";
 import {DataZoomType} from '../../option/define/datazoom';
+import { LabelOptionConverter } from '../../option/converter/label-option-converter';
+import { TooltipOptionConverter } from '../../option/converter/tooltip-option-converter';
 
 const transparentSymbolImage: string = 'image://' + window.location.origin + '/assets/images/icon_transparent_symbol.png';
 
@@ -71,6 +73,9 @@ export class LineChartComponent extends BaseChart implements OnInit, AfterViewIn
 
   // 스플릿 옵션
   private split: LineChartSplit = new LineChartSplit();
+
+  // set previous pivot (compare previous pivot, current pivot)
+  private prevPivot: Pivot;
 
   /*-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
    | Protected Variables
@@ -311,49 +316,114 @@ export class LineChartComponent extends BaseChart implements OnInit, AfterViewIn
   }
 
   /**
-   * dataLabel, tooltip 중첩에 따라서 설정
+   * change dataLabel, tooltip by single series, multi series
    * @returns {UIOption}
    */
   protected setDataLabel(): UIOption {
 
     if (!this.pivot || !this.pivot.aggregations || !this.pivot.rows) return this.uiOption;
 
+    /**
+     * check multi series <=> single series
+     * @type {() => boolean}
+     */
+    const checkChangeSeries = ((): boolean => {
+
+      // prev series is multi(true) or single
+      const prevSeriesMulti: boolean = this.prevPivot.aggregations.length > 1 || this.prevPivot.rows.length > 0 ? true : false;
+
+      // current series is multi(true) or single
+      const currentSeriesMulti: boolean = this.pivot.aggregations.length > 1 || this.pivot.rows.length > 0 ? true : false;
+
+      // if it's changed
+      if (prevSeriesMulti !== currentSeriesMulti) {
+
+        return true;
+      }
+
+      // not changed
+      return false;
+    });
+
     // 시리즈관련 리스트 제거
-    const spliceSeriesTypeList = ((seriesTypeList, dataLabel: any): UIOption => {
+    const spliceSeriesTypeList = ((seriesTypeList, dataLabel: any): any => {
 
-      // 미리보기 리스트가 빈값인지 체크
-      const previewFl = !!dataLabel.previewList;
-
+      // displayTypes를 찾는 index
       let index: number;
       for (const item of seriesTypeList) {
-
         index = dataLabel.displayTypes.indexOf(item);
+
         if (-1 !== index) {
           // 라벨에서 제거
           dataLabel.displayTypes[index] = null;
-
-          // previewList가 있는경우
-          if (previewFl) {
-            // 미리보기리스트에서 제거
-            _.remove(dataLabel.previewList, {value: item});
-          }
         }
       }
-      return dataLabel;
+      return dataLabel.displayTypes;
     });
 
-    // 교차선반에 dimension이나 measure의 개수가 2개이상이 아닌경우
-    if (this.pivot.aggregations.length < 2 && this.pivot.rows.length < 1) {
+    const setDefaultDisplayTypes = ((value): any => {
 
-      // 조건을 만족시 제거될 series타입 리스트
-      const seriesTypeList = [UIChartDataLabelDisplayType.SERIES_NAME, UIChartDataLabelDisplayType.SERIES_VALUE, UIChartDataLabelDisplayType.SERIES_PERCENT];
+      if (!value || !value.displayTypes) return [];
 
-      // 데이터라벨에서 series관련 설정제거
-      if (this.uiOption.dataLabel && this.uiOption.dataLabel.displayTypes) this.uiOption.dataLabel = spliceSeriesTypeList(seriesTypeList, this.uiOption.dataLabel);
+      let defaultDisplayTypes = [];
 
-      // 툴팁의 series관련 설정제거
-      if (this.uiOption.toolTip && this.uiOption.toolTip.displayTypes) this.uiOption.toolTip = spliceSeriesTypeList(seriesTypeList, this.uiOption.toolTip);
+      // when it has single series
+      if (this.pivot.aggregations.length <= 1 && this.pivot.rows.length < 1) {
+
+        // set disabled list when it has single series
+        const disabledList = [UIChartDataLabelDisplayType.SERIES_NAME, UIChartDataLabelDisplayType.SERIES_VALUE, UIChartDataLabelDisplayType.SERIES_PERCENT];
+
+        // remove disabled list
+        defaultDisplayTypes = spliceSeriesTypeList(disabledList, value);
+
+        // set default datalabel, tooltip list
+        defaultDisplayTypes[0] = UIChartDataLabelDisplayType.CATEGORY_NAME;
+        defaultDisplayTypes[1] = UIChartDataLabelDisplayType.CATEGORY_VALUE;
+        // when it has multi series
+      } else {
+
+        // set disabled list when it has multi series
+        const disabledList = [UIChartDataLabelDisplayType.CATEGORY_VALUE, UIChartDataLabelDisplayType.CATEGORY_PERCENT];
+
+        // remove disabled list
+        defaultDisplayTypes = spliceSeriesTypeList(disabledList, value);
+
+        // set default datalabel, tooltip list
+        defaultDisplayTypes[3] = UIChartDataLabelDisplayType.SERIES_NAME;
+        defaultDisplayTypes[4] = UIChartDataLabelDisplayType.SERIES_VALUE;
+      }
+
+      return defaultDisplayTypes;
+    });
+
+    // when draw chart or change single <=> multi series
+    if ((EventType.CHANGE_PIVOT === this.drawByType && (!this.prevPivot || checkChangeSeries())) || EventType.CHART_TYPE === this.drawByType) {
+
+      // set datalabel display types
+      let datalabelDisplayTypes = setDefaultDisplayTypes(this.uiOption.dataLabel);
+
+      // set tooltip display types
+      let tooltipDisplayTypes = setDefaultDisplayTypes(this.uiOption.toolTip);
+
+      // set default datalabel value
+      if (this.uiOption.dataLabel && this.uiOption.dataLabel.displayTypes) {
+        // set dataLabel
+        this.uiOption.dataLabel.displayTypes = datalabelDisplayTypes;
+        // set previewList
+        this.uiOption.dataLabel.previewList = LabelOptionConverter.setDataLabelPreviewList(this.uiOption);
+      }
+
+      // set default tooltip value
+      if (this.uiOption.toolTip && this.uiOption.toolTip.displayTypes) {
+        // set dataLabel
+        this.uiOption.toolTip.displayTypes = tooltipDisplayTypes;
+        // set previewList
+        this.uiOption.toolTip.previewList = TooltipOptionConverter.setTooltipPreviewList(this.uiOption);
+      }
     }
+
+    // set previous pivot value (compare previous pivot, current pivot)
+    this.prevPivot = this.pivot;
 
     return this.uiOption;
   }
